+++ priv/repo/migrations/20260729142054_imports_add_file.exs
@@ -1,0 +1,9 @@
+defmodule Linkhut.Repo.Migrations.ImportsAddFile do
+  use Ecto.Migration
+
+  def change do
+    alter table(:imports) do
+      add :file, :string
+    end
+  end
+end